+++ warc-indexer/src/main/java/uk/bl/wa/analyser/payload/WARCPayloadAnalysers.java
@@ -70,7 +70,6 @@ public class WARCPayloadAnalysers {
 	public PDFAnalyser pdf;
 	public XMLAnalyser xml;
 	public ImageAnalyser image;
-    public ARCNameAnalyser arcname;
 
 	private boolean extractApachePreflightErrors;
 	private boolean extractImageFeatures;
@@ -109,8 +108,7 @@ public class WARCPayloadAnalysers {
 		if (this.extractImageFeatures) {
 			image = new ImageAnalyser(conf);
 		}
-        arcname = new ARCNameAnalyser(conf);
-		Instrument.createSortedStat("WARCPayloadAnalyzers.analyze#droid", Instrument.SORT.avgtime, 5);
+        Instrument.createSortedStat("WARCPayloadAnalyzers.analyze#droid", Instrument.SORT.avgtime, 5);
 	}
 	
 	public void analyse(ArchiveRecordHeader header, InputStream tikainput, SolrRecord solr) {
@@ -183,14 +181,7 @@ public class WARCPayloadAnalysers {
 
 		}
 
-        // Parse ARC name
-        if (!arcname.getRules().isEmpty()) {
-            final long nameStart = System.nanoTime();
-            arcname.analyse(header, tikainput, solr);
-            Instrument.timeRel("WARCPayloadAnalyzers.analyze#total",
-                               "WARCPayloadAnalyzers.analyze#arcname", nameStart);
-        }
-
+     
 		try {
 			tikainput.reset();
 			String mime = ( String ) solr.getField( SolrFields.SOLR_CONTENT_TYPE ).getValue();